+++ packages/cli/lib/util/getInstalledVersions.js
@@ -1,7 +1,10 @@
 import fs from 'node:fs/promises'
-import { versionRegEx } from './constants.js'
-import { TfvmFS } from './getDirectoriesObj.js'
+import { versionRegEx, LOWEST_OTF_VERSION } from './constants.js'
 import { logger } from './logger.js'
+import getSettings from './getSettings.js'
+import { getOS } from './tfvmOS.js'
+import * as semver from 'semver'
+const os = getOS()
 
 let installedVersions
 
@@ -9,21 +12,42 @@ let installedVersions
  * Returns a list of installed tf versions.
  * @returns {Promise<string[]>}
  */
-async function getInstalledVersions () {
+async function getInstalledVersions (version = '') {
+  const settings = await getSettings()
   // return the list of installed versions if that is already cached
   if (!installedVersions) {
-    const tfList = []
+    const versionsList = []
+    let files
 
-    const files = await fs.readdir(TfvmFS.tfVersionsDir)
+    let semverCheck = true
+    if (version !== '') {
+      semverCheck = semver.gte(version, LOWEST_OTF_VERSION)
+    }
+    if (settings.useOpenTofu && semverCheck) {
+      files = await fs.readdir(os.getOtfVersionsDir())
+      const terraformFiles = await fs.readdir(os.getTfVersionsDir())
+      terraformFiles.forEach(file => {
+        if (semver.lt(file, LOWEST_OTF_VERSION)) {
+          files.push(file)
+        }
+      })
+    } else {
+      files = await fs.readdir(os.getTfVersionsDir())
+    }
+
     if (files && files.length) {
       files.forEach(file => {
         if (versionRegEx.test(file)) {
-          tfList.push(file)
+          versionsList.push(file)
         }
       })
-      installedVersions = tfList
+      installedVersions = versionsList
     } else {
-      logger.debug(`Unable to find installed versions of terraform with directoriesObj=${JSON.stringify(TfvmFS.getDirectoriesObj())} and files=${JSON.stringify(files)}`)
+      if (settings.useOpenTofu) {
+        logger.debug(`Unable to find installed versions of OpenTofu with directoriesObj=${JSON.stringify(os.getDirectories())} and files=${JSON.stringify(files)}`)
+      } else {
+        logger.debug(`Unable to find installed versions of Terraform with directoriesObj=${JSON.stringify(os.getDirectories())} and files=${JSON.stringify(files)}`)
+      }
       return []
     }
   }