+++ packages/cli/lib/util/tfVersion.js
@@ -1,22 +1,40 @@
 import runShell from './runShell.js'
 import { logger } from './logger.js'
-import { tfCurrVersionRegEx } from './constants.js'
+import { tfCurrVersionRegEx, openTofuCurrVersionRegEx } from './constants.js'
+import getSettings from './getSettings.js'
 
 let currentTfVersion
+let currentOtfVersion
 
 /**
  * Returns the current terraform version, if there is one. Returns null if there is no current version
  * @returns {Promise<string|null>}
  */
 async function getTerraformVersion () {
+  const settings = await getSettings()
   // cache current tf version during a single execution of tfvm
-  if (currentTfVersion) return currentTfVersion
-  const response = (await runShell('terraform -v'))
+  if (currentTfVersion && !settings.useOpenTofu) return currentTfVersion
+  if (currentOtfVersion && settings.useOpenTofu) return currentOtfVersion
+
+  let response
+  if (settings.useOpenTofu) {
+    response = (await runShell('tofu -v'))
+    if (response === null) {
+      response = (await runShell('terraform -v'))
+    }
+  } else {
+    response = (await runShell('terraform -v'))
+  }
+
   if (response == null) {
     logger.error('Error getting terraform version')
     return null
   }
-  const versionExtractionResult = Array.from(response.matchAll(tfCurrVersionRegEx))
+
+  let versionExtractionResult
+  if (response.includes('Terraform')) versionExtractionResult = Array.from(response.matchAll(tfCurrVersionRegEx))
+  else versionExtractionResult = Array.from(response.matchAll(openTofuCurrVersionRegEx))
+
   if (versionExtractionResult.length === 0) {
     logger.error('Error extracting terraform version where this is the response from `terraform -v`:\n' + response)
     return null